// lib/cli/generated/completions.ts
export const completionMeta: CompletionMeta = {
  valueFlags: {
    'universe': 'universe',
    'project': 'project',
  },
  commands: {
// ...
    'list': [{name: 'project', kind: 'project', variadic: false}, {name: 'substring', kind: null, variadic: false}],
    'delete': [{name: 'project', kind: 'project', variadic: false}, {name: 'target', kind: null, variadic: true}],
    'flamegraph': [{name: 'project', kind: 'project', variadic: false}],
    'symbold status': [{name: 'project', kind: 'project', variadic: false}],
    'symbold symbolserver list': [{name: 'project', kind: 'project', variadic: false}],
    'symbold symbolserver add': [{name: 'project', kind: 'project', variadic: false}, {name: 'url', kind: null, variadic: false}],
    'symbold queue add': [{name: 'project', kind: 'project', variadic: false}, {name: 'symbol', kind: null, variadic: false}, {name: 'oid', kind: null, variadic: false}],
    'report list': [{name: 'project', kind: 'project', variadic: false}],
    'report create': [{name: 'project', kind: 'project', variadic: false}],
    'report delete': [{name: 'project', kind: 'project', variadic: false}, {name: 'id', kind: null, variadic: false}],
    'report send': [{name: 'project', kind: 'project', variadic: false}, {name: 'id', kind: null, variadic: false}, {name: 'email', kind: null, variadic: false}],
    'merge': [{name: 'project', kind: 'project', variadic: false}, {name: 'fingerprints', kind: null, variadic: true}],
    'unmerge': [{name: 'project', kind: 'project', variadic: false}, {name: 'fingerprints', kind: null, variadic: true}],
    'repair': [{name: 'project', kind: 'project', variadic: false}],
    'reprocess': [{name: 'project', kind: 'project', variadic: false}, {name: 'target', kind: null, variadic: true}],
    'symbol': [{name: 'project', kind: 'project', variadic: false}, {name: 'action', kind: null, variadic: false}],
    'scrubber list': [{name: 'project', kind: 'project', variadic: false}],
    'scrubber create': [{name: 'project', kind: 'project', variadic: false}],
    'scrubber modify': [{name: 'project', kind: 'project', variadic: false}, {name: 'id', kind: null, variadic: false}],
    'scrubber delete': [{name: 'project', kind: 'project', variadic: false}, {name: 'id', kind: null, variadic: false}],
    'similarity': [{name: 'project', kind: 'project', variadic: false}, {name: 'filter', kind: null, variadic: false}],
    'callstack evaluate': [{name: 'project', kind: 'project', variadic: false}, {name: 'target', kind: null, variadic: false}],
    'access project add-team': [{name: 'project', kind: 'project', variadic: false}, {name: 'team', kind: null, variadic: false}, {name: 'role', kind: null, variadic: false}],
    'access project remove-team': [{name: 'project', kind: 'project', variadic: false}, {name: 'team', kind: null, variadic: false}],
    'access project add-user': [{name: 'project', kind: 'project', variadic: false}, {name: 'user', kind: null, variadic: false}, {name: 'role', kind: null, variadic: false}],
    'access project remove-user': [{name: 'project', kind: 'project', variadic: false}, {name: 'user', kind: null, variadic: false}],
    'access project details': [{name: 'project', kind: 'project', variadic: false}],
    'attribute create': [{name: 'project', kind: 'project', variadic: false}, {name: 'name', kind: null, variadic: false}],
    'attribute delete': [{name: 'project', kind: 'project', variadic: false}, {name: 'name', kind: null, variadic: false}],
    'view create': [{name: 'project', kind: 'project', variadic: false}, {name: 'name', kind: null, variadic: false}],
    'view delete': [{name: 'project', kind: 'project', variadic: false}, {name: 'name', kind: null, variadic: false}],
    'deduplication add': [{name: 'project', kind: 'project', variadic: false}],
    'deduplication delete': [{name: 'project', kind: 'project', variadic: false}],
    'deduplication modify': [{name: 'project', kind: 'project', variadic: false}],
    'deduplication list': [{name: 'project', kind: 'project', variadic: false}],
    'cts': [{name: 'project', kind: 'project', variadic: false}, {name: 'attribute', kind: null, variadic: false}, {name: 'value', kind: null, variadic: false}],
  },
};